% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{combatrefql-package}
\alias{combatrefql}
\alias{combatrefql-package}
\title{ComBat-refQL: reference-batch adjustment for RNA-seq counts}
\description{
A structured interface using edgeR v4 model fitting, evidence-adaptive EB
batch contrasts, hierarchical batch-specific NB dispersion, deterministic
mid-P transport, and validated S7 diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thokas99.github.io/ComBat-refQL/}
  \item \url{https://github.com/Thokas99/ComBat-refQL}
  \item Report bugs at \url{https://github.com/Thokas99/ComBat-refQL/issues}
}

}
\author{
\strong{Maintainer}: Thomas Sirchi \email{thomasirchi@gmail.com}

Authors:
\itemize{
  \item Thomas Sirchi \email{thomasirchi@gmail.com}
}

Other contributors:
\itemize{
  \item Xiaoyu Zhang (ComBat-ref method and original implementation) [contributor]
}

}
\keyword{internal}
